+++ CHANGELOG.md
@@ -3,12 +3,13 @@ Changelog
 
 For detailed information check [Blogophon's releases on Github](https://github.com/fboes/blogophon/releases).
 
-rc 1.1.0
---------
+1.1.0
+-----
 
 * [x] New (and very basic) default theme
-* [x] Fixed major bug for installations in subfolders
 * [x] Image gallery added
+* [x] Adding support for video / audio tags in Markdown
+* [x] Fixed major bug for installations in subfolders
 
 1.0.4
 -----